move all pages
