+++ src/pages/api/juicebox/pv/[pv]/project/[projectId]/logo.ts
