+++ src/pages/legal/index.tsx
